skip benchmark with GNU error
