+++ benchmarks/benchmark_base.py
@@ -394,3 +394,22 @@ class BenchmarkBase:
             v_inner=np.array([-1, -1], dtype=np.float64),
             v_outer=np.array([-1, -1], dtype=np.float64),
         )
+
+
+    @property
+    def estimators(self):
+        return radfield_mc_estimators.RadiationFieldMCEstimators(
+            j_estimator=np.array([0.0, 0.0], dtype=np.float64),
+            nu_bar_estimator=np.array([0.0, 0.0], dtype=np.float64),
+            j_blue_estimator=np.array(
+                [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0]], dtype=np.float64
+            ),
+            Edotlu_estimator=np.array(
+                [[0.0, 0.0, 1.0], [0.0, 0.0, 1.0]], dtype=np.float64
+            ),
+            photo_ion_estimator=np.empty((0, 0), dtype=np.float64),
+            stim_recomb_estimator=np.empty((0, 0), dtype=np.float64),
+            bf_heating_estimator=np.empty((0, 0), dtype=np.float64),
+            stim_recomb_cooling_estimator=np.empty((0, 0), dtype=np.float64),
+            photo_ion_estimator_statistics=np.empty((0, 0), dtype=np.int64),
+        )
