+++ benchmarks/run_tardis.py
@@ -4,8 +4,6 @@ Basic TARDIS Benchmark.
 
 from benchmarks.benchmark_base import BenchmarkBase
 from tardis import run_tardis
-from tardis.io.configuration.config_reader import Configuration
-
 
 class BenchmarkRunTardis(BenchmarkBase):
     """
@@ -14,12 +12,8 @@ class BenchmarkRunTardis(BenchmarkBase):
 
     def __init__(self):
         super().__init__()
-        self.config = None
-
-    def setup(self):
         filename = "data/tardis_configv1_benchmark.yml"
-        path = self.get_relative_path(filename)
-        self.config = Configuration.from_yaml(path)
+        self.path = self.get_relative_path(filename)
 
     def time_run_tardis(self):
-        run_tardis(self.config, log_level="ERROR", show_progress_bars=False)
+        run_tardis(self.path, log_level="ERROR", show_progress_bars=False)
