+++ benchmarks/transport_montecarlo_packet.py
@@ -23,24 +23,6 @@ class BenchmarkMontecarloMontecarloNumbaPacket(BenchmarkBase):
     Class to benchmark the numba packet function.
     """
 
-    @property
-    def estimators(self):
-        return radfield_mc_estimators.RadiationFieldMCEstimators(
-            j_estimator=np.array([0.0, 0.0], dtype=np.float64),
-            nu_bar_estimator=np.array([0.0, 0.0], dtype=np.float64),
-            j_blue_estimator=np.array(
-                [[0.0, 0.0, 0.0], [0.0, 0.0, 0.0]], dtype=np.float64
-            ),
-            Edotlu_estimator=np.array(
-                [[0.0, 0.0, 1.0], [0.0, 0.0, 1.0]], dtype=np.float64
-            ),
-            photo_ion_estimator=np.empty((0, 0), dtype=np.float64),
-            stim_recomb_estimator=np.empty((0, 0), dtype=np.float64),
-            bf_heating_estimator=np.empty((0, 0), dtype=np.float64),
-            stim_recomb_cooling_estimator=np.empty((0, 0), dtype=np.float64),
-            photo_ion_estimator_statistics=np.empty((0, 0), dtype=np.int64),
-        )
-
     @parameterize(
         {
             "Parameters": [
